Improve parsing of arguments
