+++ src/compiler/util/parseJavaScriptParams.js
@@ -1,20 +1,23 @@
 "use strict";
 
-var ok = require("assert").ok;
-var esprima = require("esprima");
+const ok = require("assert").ok;
+const parseRawJavaScriptAst = require("./parseRawJavaScriptAst");
 
 function parseJavaScriptParams(params, builder) {
     ok(typeof params === "string", '"params" should be a string');
     ok(builder, '"builder" is required');
 
-    var src = "(" + params + ") => {}";
-    var jsAST = esprima.parseScript(src, { range: true });
-    var paramNodes = jsAST.body[0].expression.params;
-    return paramNodes.map(node => {
-        var nodeSrc = src.slice(node.range[0], node.range[1]);
+    const src = "(" + params + ") => {}";
+    const ast = parseRawJavaScriptAst(src, {
+        startOffset: 1,
+        endOffset: -7
+    });
+
+    return ast.body[0].expression.params.map(node => {
+        const paramSrc = src.slice(node.range[0], node.range[1]);
         return node.type === "Identifier"
-            ? builder.identifier(nodeSrc)
-            : builder.expression(nodeSrc);
+            ? builder.identifier(paramSrc)
+            : builder.expression(paramSrc);
     });
 }
