+++ test/compiler/fixtures-vdom/arrow-function-event-handler/expected.js
@@ -1,0 +1,25 @@
+"use strict";
+
+var marko_template = module.exports = require("marko/src/vdom").t(),
+    components_helpers = require("marko/src/components/helpers"),
+    marko_registerComponent = components_helpers.rc,
+    marko_componentType = marko_registerComponent("/marko-test$1.0.0/compiler/fixtures-vdom/arrow-function-event-handler/template.marko", function() {
+      return module.exports;
+    }),
+    marko_renderer = components_helpers.r,
+    marko_defineComponent = components_helpers.c;
+
+function render(input, out, __component, component, state) {
+  var data = input;
+
+  out.e("DIV", null, "0", component, 0, 0, {
+      onclick: __component.d("click", () => console.log("handled"), false)
+    });
+}
+
+marko_template._ = marko_renderer(render, {
+    ___implicit: true,
+    ___type: marko_componentType
+  });
+
+marko_template.Component = marko_defineComponent({}, marko_template._);
