+++ test/compiler/fixtures-vdom/arrow-function-event-handler/template.marko
@@ -1,0 +1,1 @@
+<div on-click(() => console.log("handled"))/>
